refactor(sliderCaptcha): extract config and hooks from SliderCaptchaContent
- config/sliderCaptchaTypes.ts: SliderCaptchaContentProps interface
- hooks/useSliderCaptchaDerived.ts: slider progress, style, expression, traceCode
- utils/.gitkeep, components/.gitkeep: empty directory placeholders
- SliderCaptchaContent.tsx: simplified from 150 to ~110 lines
